MAJOR COMMIT
Made insane numbers of changes. Had huge issue sending packets because
one out of ever 3 packets would be corrupt and it would ruin the whole
program. So I made it so multiple packets send for each packet. For
example I would send 5 packets and they will all have the same UUID, and
hopefully none of them are ever corrupt because the chance is 5 times
less (maybe i should make it even more than 5..) but basically when a
PacketReceiver (new class) receives a packet that is valid, it will add
it to a HashMap that is essentially a packet queue, but they are sorted
in ascending order based on the packets regular ID. It will NOT add it
to the queue if there is already a packet in there with the same UUID.
This new abstraction that I made for packet receiving (PacketReceiver,
PacketListener classes) is quite awesome (if it works which it might
not). It makes passing the socket from ConnectionGUI -> LobbyGUI ->
GameGUI a lot easier, and gives me peace of mind about receiving packets
in the right order and receiving them properly