HUGE BUGFIX
I made the getUTF() method in packet not contain the UUID after I added
the UUID. I just made the send method do <UUID> + getUTF(). AND I LEFT
THIS HERE BECAUSE I USED IT TO MAKE SURE THAT THE PACKETS THAT WERE
BEING SENT WERE VALID BECAUSE I WAS SURE THEY WERE (THEY WERE, BUT THIS
WAS BEFORE I ADDED UUID). BUT I FORGOT TO UPDATE IT WITH THE UUID METHOD
AND IT MADE EVERYTHING REALLY DIFFICULT FOR ME