SUPER IMPORTANT COMMIT - MAJOR BUG FIXES
I can finally safely say that, as far as I know, I have fixed the huge
bug I was getting with transferring packets between sockets.
- Completely revamped the server/client socket programming and made it
easier to transfer the client's interface with the server over to other
GUIs. The solution was to handle the client's packet receiving in ONE
class, called Client (the class that used to be called client is now
called SnakeClient). When the Client class receives packets, it sends
them to its connected ClientListener field (which is an Interface). With
this setup I can just make all the different Client GUI's implement
ClientListener and pass the client object between them, and once I pass
the client object to the next GUI, I can type client.setListener(new
GUI).
- Reverted packets back to how they originally were. For some reason I
thought packets were sending out of order and causing all of the
crashing but that was just because I coded it like spaghetti. To be
honest, I don't even know what caused the bug in the first place. All I
know is that when I recoded the socket programming and packets, it was
fixed.
- The server now sends packets using Server.send() and
Server.sendToAll() AND IT SHOULD ONLY USE THESE METHODS, NEVER DIRECTLY
USE Packet.send. This is because the server is multi-threaded and
listening from many clients, and when a client sends a packet to the
server, the server is pretty much always going to bounce that packet to
all the clients. So to avoid any future issues of 2 clients sending
packets at the same time and consequentially causing the server to call
Packet.send() twice at the same time, I made the Server.send() and
Server.sendToAll() methods synchronized