Added new features
- Renamed BuffDrawScript to SnakeDrawScript because I figured I could
use BuffDrawScript for tons of other animations that have to do with the
snake but not necessarily buff animations
- Added a new type of Packet: SnakeDeathPacket. When a client's snake
dies the server will send a SnakeDeathPacket containing the dead
client's name. When it is received by a client, a DeadSnakeDrawScript
will be added to the dead client's SnakeData.
- Added a new type of SnakeDrawScript: DeadSnakeDrawScript. This is
triggered on the client-side for a specific user's SnakeData. It is used
to play a death animation for the snake, which is an explosion.
- Added a new class: Vector. This is mainly used by the
DeadSnakeDrawScript for the explosions