MultiThreading and Backup
+ Added ThreadTool class which has a start index and end index (indicies
in a very large array) and a static method that makes a list of
ThreadTools based on the size of the list and how many Threads you want
to split the list into.
+ Improved efficiency of doSnakeMovements() in the server.
+ Fixed Lag issues with the HungrySnakeDrawScript

* This commit also serves as a backup. I am going to try to change the
way SnakeUpdatePacket works to make it more efficient for the client
when the Snakes are really long.