Huge Commit
+ Added SnakeBiteDrawScript and SnakeBitePacket. These both have to do
with the HungryBuffDrawScript and the Hungry Buff. When a user has the
Hungry buff they can bite off some segments from the end of other snakes
and steal them. The maximum amount they can bite off currently is 30. If
you try to bite off any more than that, or the snake you are biting from
isn't 31 or greater units long, then you will die. Biting off segments
from a snake also removes whatever buff they currently have.
+ Changed the way SnakeUpdatePackets are sent. After the SnakeLists on
the client and server are initialized, the server will stop sending the
PointLists to the client. This is because it gets extremely inefficient
to send super long point lists to the clients to update the snakes.
Instead the clients can use a simple 3 step method of updating the point
lists on their own.
* Might potentially remove some of the multithreading in some of the
SnakeDrawScript subclasses as it may be unncecessary.
* May consider painting each snake on a different Thread in
GameplayClient.java
* TODO: I want to make it so there are two different types of
HungryBuffDrawScript. One type will be created for the Snake that is
currently hungry and that is what their client will display. It will be
different in the sense that it highlights the segments of enemy snakes
that it is able to bite off. For the clients of the Snake's who AREN'T
hungry, they will receive a different DrawScript where these segments
aren't highlighted