Lots of changes
+Found out that the reason why I needed to convert the PointList to an
Array before drawing it in HungryBuffDrawScript was because my
getPointList() method always returns a copy of the list, which makes
drawing this script take O(n^2) time to complete. When the snake has
100+ units, it lags the program a lot. So instead of converting to an
Array, I made the getPointList() method have a boolean on whether or not
to clone it. Any time the point list is changed, a clone is first made,
then changed, then the PointList is set to the clone. This is to avoid
ConcurrentModificationExceptions for the multi-threaded painting of some
scripts.
+ HungryBuffDrawScript now highlights the parts of enemies that can be
bitten off for the player who has the buff
+ Other changes