Cleanup, Improvements, Bug fixes, Documentation
+Concurrent thread issues when receive() and send() are called at the
same time have been fixed as far as I know. Basically, receive() and
send() methods in all relevant classes (such as GameplayClient,
LobbyClient etc) should be synchronized together with their class as the
lock so the classe's SocketSecurityTool isn't encrypting/decrypting
anything concurrently. Concurrent encryption/decryption can cause issues
with SocketSecurityTool because they both call the init method on the
cipher and its possible that while one method is trying to encrypt, the
cipher could be set to decrypt mode by the other method which will cause
a BadPaddingException. I also added synchronization in the
SocketSecurityTool as an extra precaution against this. Even with this
extra precaution, send() and receive() methods should be synchronized
because you can run into weird socket writing errors if two sends are
called concurrently. On top of this, if send and receive are called
concurrently there are many other issues that you can run into besides
just the SocketSecurityTool errors.

+The LobbyGUI client is no longer in control of figuring out what
PlayerSlot (renamed from PlayerDisplayPanel) to put each client in. It
is also no longer responsible for determining the start position of the
Snake's and their colors. This is all up to the server now.