Fixed Encryption Bugs, Backup
Fixed issue with Asymmetric encryption where the data was too long to
encrypt. I fixed this issue by segmenting the byte[] that is being sent.
However, I want to change this to where I use an AES key to encrypt the
byte[] because length doesn't matter with AES, then I will encrypt the
AES key with RSA and send the key along with the encrypted data. I am
using this commit as a backup