Bug Fixes
+ Fixed latency issues when connecting to other computers by using
Socket.setTCPNoDelay(true).
+ Replaced all DataInputStream and DataOutputStream references to
BufferedInputStream and BufferedOutputStream. This way I can write
multiple bytes to the buffer and send them in one TCP packet using
flush. This makes up for the inefficiency of turning on TCPNoDelay
(disabling Nagle's algorithm). This also fixes a huge bug where
sometimes DataInputStream.write will automatically try and buffer the
information, so it won't actually be sent until later which can mess
things up because the data needs to be sent in order